Split behaviors and items into multiple files
