+++ core/src/main/scala/io/github/scalaquest/core/model/common/items/DefaultCommonItems.scala
@@ -1,0 +1,8 @@
+package io.github.scalaquest.core.model.common.items
+
+import io.github.scalaquest.core.model.common.items.impl.{Door, GenericItem, Key}
+
+/**
+ * This is a mixable part of the model, that adds some implemented common items to the model.
+ */
+trait DefaultCommonItems extends DefaultCommonItemsBase with Door with GenericItem with Key {}
