+++ core/src/main/scala/io/github/scalaquest/core/model/common/items/DefaultCommonItemsBase.scala
@@ -1,0 +1,6 @@
+package io.github.scalaquest.core.model.common.items
+
+import io.github.scalaquest.core.model.common.behaviors.CommonBehaviors
+import io.github.scalaquest.core.model.default.BehaviorableModel
+
+trait DefaultCommonItemsBase extends BehaviorableModel with CommonBehaviors with CommonItems
