+++ core/src/main/scala/io/github/scalaquest/core/model/common/items/impl/GenericItem.scala
@@ -1,0 +1,7 @@
+package io.github.scalaquest.core.model.common.items.impl
+
+import io.github.scalaquest.core.model.common.items.DefaultCommonItemsBase
+
+trait GenericItem extends DefaultCommonItemsBase {
+  case class GenericItem(name: String, override val behaviors: Set[Behavior] = Set()) extends CommonItems.GenericItem
+}
